+++ src/Traits/Translatable.php
@@ -18,10 +18,18 @@ trait Translatable
      */
     public function translation($lang)
     {
-        return Translation::where('entity_id', $this->id)
+        $result = Translation::where('entity_id', $this->id)
             ->where('entity_type', get_class($this))
-            ->where('entity_data', 'LIKE', '%"lang":"'.$lang.'"%')
+            ->where('language', $lang)
             ->first();
+
+        if ($result) {
+            return $result;
+        }
+
+        $this->data = $this;
+
+        return $this;
     }
 
     /**
@@ -84,17 +92,17 @@ trait Translatable
 
             foreach (config('cms.languages') as $code => $language) {
                 if ($code != config('cms.default-language')) {
-                    $tr = new TranslateClient(config('cms.default-language'), $code);
+                    $translateClient = new TranslateClient(config('cms.default-language'), $code);
                     $translation = [
                         'lang' => $code,
                         'template' => 'show',
                     ];
 
                     foreach ($entry as $key => $value) {
+                        $translation[$key] = $value;
+
                         if (!empty($value)) {
-                            $translation[$key] = json_decode(json_encode($tr->translate(strip_tags($value))));
-                        } else {
-                            $translation[$key] = $value;
+                            $translation[$key] = json_decode(json_encode($translateClient->translate(strip_tags($value))));
                         }
                     }