+++ src/Views/modules/images/index.blade.php
@@ -43,8 +43,8 @@
     <div class="col-md-12">
         <nav class="navbar px-0 navbar-light justify-content-between">
             <div class="navbar-nav navbar-expand-md mr-auto justify-content-between">
-                <a class="d-inline nav-item btn btn-primary mr-1 mt-2" href="{!! route(cms()->route('images.create')) !!}">Add New</a>
-                <button class="d-inline nav-item btn btn-danger bulk-image-delete mt-2"><span class="fa fa-trash"></span> Delete</button>
+                <a class="nav-item btn btn-primary mr-1 mt-2" href="{!! route(cms()->route('images.create')) !!}">Add New</a>
+                <button class="nav-item btn btn-danger bulk-image-delete mt-2"><span class="fa fa-trash"></span> Bulk Delete</button>
             </div>
             {!! Form::open(['url' => cms()->url('images/search'), 'class' => 'form-inline mt-2']) !!}
                 <input class="form-control mr-sm-2" name="term" type="search" placeholder="Search" aria-label="Search">
@@ -71,14 +71,14 @@
                                 </div>
                                 <div data-id="{{ $image->id }}" class="well pull-down overflow-hidden selectable">
                                     <div class="row">
-                                        <div class="col-lg-6 col-md-12 col-sm-12">
+                                        <div class="col-6">
                                             @if ($image->is_published)
                                                 <span clas="pull-left"><span class="pull-left fa fa-check"></span> Published</span>
                                             @else
                                                 <span clas="pull-left"><span class="pull-left fa fa-times"></span> Published</span>
                                             @endif
                                         </div>
-                                        <div class="col-lg-6 col-md-12 col-sm-12">
+                                        <div class="col-6">
                                             <div class="btn-toolbar float-right">
                                                 <a class="btn btn-sm btn-secondary img-alter-btn mr-2" href="{!! route(cms()->route('images.edit'), [$image->id]) !!}"><i class="fa fa-edit"></i></a>
                                                 <form method="post" action="{!! cms()->url('images/'.$image->id) !!}">
@@ -103,3 +103,10 @@
     </div>
 
 @endsection
+
+@section('pre_javascript')
+
+    @parent
+    var _cmsUrl = "{{ cms()->url('/') }}";
+
+@stop